refactor: rename Technical and Skills components
This commit renames the following components and folders:

1. Renamed `Technical.jsx` to `Skills.jsx` to better reflect the purpose of the component.
2. Renamed `Skills.jsx` to `Workflow.jsx` to provide a more accurate representation of the component's functionality.

These changes ensure better alignment between component names and their actual content, improving code clarity and maintainability.
