--- california-dev/src/components/sections/Technical/index.js
@@ -1,1 +1,0 @@
-export * from "./Technical";
