+++ california-dev/src/components/sections/WorkFlow/index.js
@@ -1,0 +1,1 @@
+export * from "./Workflow";
